requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pillow>=10.0.0
matplotlib>=3.7.0
scipy>=1.10.0

# COCO API and detection
pycocotools>=2.0.6

# Data handling
pandas>=2.0.0
opencv-python>=4.8.0

# Visualization
seaborn>=0.12.0
tqdm>=4.65.0

# Configuration
pyyaml>=6.0

# Experiment tracking (optional)
tensorboard>=2.13.0
